@@ -89,8 +86,8 @@
     </dependencies>
 
     <properties>
-        <maven.compiler.source>13</maven.compiler.source>
-        <maven.compiler.target>13</maven.compiler.target>
+        <maven.compiler.source>14</maven.compiler.source>
+        <maven.compiler.target>14</maven.compiler.target>
         <junit.jupiter.version>5.3.1</junit.jupiter.version>
         <junit.platform.version>1.3.1</junit.platform.version>
     </properties>
@@ -152,14 +149,9 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.0</version>
                 <configuration>
-                    <source>13</source>
-                    <target>13</target>
+                    <source>14</source>
+                    <target>14</target>
                     <verbose>true</verbose>
-                    <compilerArgs>
-                        <arg>
-                            --enable-preview
-                        </arg>
-                    </compilerArgs>
                 </configuration>
             </plugin>
             <plugin>
