@@ -4,17 +4,17 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public enum SNESRomSpeed {
-    SLOWROM, FASTROM;
+  SLOWROM,
+  FASTROM;
 
-
-    /**
-     * Gets these values as a List of strings for text matching
-     * @return a list of these values, stringified
-     */
-    public static List<String> asCollection() {
-        return List.of(SNESRomSpeed.values())
-                .stream()
-                .map(SNESRomSpeed::name)
-                .collect(Collectors.toList());
-    }
+  /**
+   * Gets these values as a List of strings for text matching
+   *
+   * @return a list of these values, stringified
+   */
+  public static List<String> asCollection() {
+    return List.of(SNESRomSpeed.values()).stream()
+        .map(SNESRomSpeed::name)
+        .collect(Collectors.toList());
+  }
 }
