@@ -1,9 +1,5 @@
 package dev.secondsun.wla4j.assembler.analyzer;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import dev.secondsun.wla4j.assembler.definition.directives.AllDirectives;
 import dev.secondsun.wla4j.assembler.pass.parse.ErrorNode;
 import dev.secondsun.wla4j.assembler.pass.parse.Node;
@@ -11,6 +7,10 @@ import dev.secondsun.wla4j.assembler.pass.parse.ParseException;
 import dev.secondsun.wla4j.assembler.pass.parse.SlotNode;
 import dev.secondsun.wla4j.assembler.pass.parse.directive.DirectiveNode;
 import dev.secondsun.wla4j.assembler.pass.parse.directive.definition.DefinitionNode;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class MemoryMapAnalyzer extends AbstractAnalyzer {
 
