@@ -1,10 +1,10 @@
 package dev.secondsun.wla4j.assembler.analyzer;
 
-import java.util.ArrayList;
-import java.util.List;
 import dev.secondsun.wla4j.assembler.pass.parse.ErrorNode;
 import dev.secondsun.wla4j.assembler.pass.parse.ParseException;
 import dev.secondsun.wla4j.assembler.pass.parse.directive.DirectiveNode;
+import java.util.ArrayList;
+import java.util.List;
 
 public class SNESNativeVectorAnalyzer extends AbstractAnalyzer {
 
@@ -17,13 +17,17 @@ public class SNESNativeVectorAnalyzer extends AbstractAnalyzer {
     var errors = new ArrayList<ErrorNode>();
 
     if (context.getSnesRomMode() == null) {
-      errors.add(new ErrorNode(node.getSourceToken(),
-          new ParseException("Rom mode must be set before SNESNATIVEVECTOR",
-              node.getSourceToken())));
+      errors.add(
+          new ErrorNode(
+              node.getSourceToken(),
+              new ParseException(
+                  "Rom mode must be set before SNESNATIVEVECTOR", node.getSourceToken())));
     }
     if (context.getSnesNativeVector()) {
-      errors.add(new ErrorNode(node.getSourceToken(),
-          new ParseException("SNESNATIVEVECTOR may only be set once", node.getSourceToken())));
+      errors.add(
+          new ErrorNode(
+              node.getSourceToken(),
+              new ParseException("SNESNATIVEVECTOR may only be set once", node.getSourceToken())));
     } else {
       context.setSnesNativeVector(true);
     }
