@@ -3,10 +3,10 @@ package dev.secondsun.wla4j.assembler.pass.parse.directive.definition;
 import static dev.secondsun.wla4j.assembler.definition.directives.AllDirectives.DS;
 import static dev.secondsun.wla4j.assembler.definition.directives.AllDirectives.DSB;
 
-import java.util.Arrays;
-import java.util.List;
 import dev.secondsun.wla4j.assembler.definition.directives.AllDirectives;
 import dev.secondsun.wla4j.assembler.pass.parse.directive.GenericDirectiveParser;
+import java.util.Arrays;
+import java.util.List;
 
 public class DefineByteSeriesParser extends GenericDirectiveParser {
 
