@@ -9,8 +9,8 @@ import dev.secondsun.wla4j.assembler.definition.directives.AllDirectives;
 import dev.secondsun.wla4j.assembler.pass.parse.NodeTypes;
 import dev.secondsun.wla4j.assembler.pass.parse.ParseException;
 import dev.secondsun.wla4j.assembler.pass.parse.SourceParser;
-import dev.secondsun.wla4j.assembler.pass.parse.directive.DirectiveBodyNode;
 import dev.secondsun.wla4j.assembler.pass.parse.directive.DirectiveArgumentsNode;
+import dev.secondsun.wla4j.assembler.pass.parse.directive.DirectiveBodyNode;
 import dev.secondsun.wla4j.assembler.pass.parse.directive.DirectiveNode;
 import dev.secondsun.wla4j.assembler.pass.parse.directive.DirectiveParser;
 import dev.secondsun.wla4j.assembler.pass.scan.token.Token;
