@@ -1,11 +1,11 @@
 package dev.secondsun.wla4j.assembler.pass.parse.visitor;
 
+import dev.secondsun.wla4j.assembler.pass.parse.Node;
+import dev.secondsun.wla4j.assembler.pass.parse.directive.macro.MacroNode;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
-import dev.secondsun.wla4j.assembler.pass.parse.Node;
-import dev.secondsun.wla4j.assembler.pass.parse.directive.macro.MacroNode;
 
 /** When a macro is defined, add the relevant information to a Map. */
 public class MacroDefinitionVisitor implements Visitor {
