@@ -4,16 +4,15 @@ import dev.secondsun.wla4j.assembler.analyzer.Context;
 import dev.secondsun.wla4j.assembler.main.Project;
 
 public class SourceCompiler {
-    private final Project project;
-    private final Context context;
+  private final Project project;
+  private final Context context;
 
-    public SourceCompiler(Project helloWorldProject, Context ctx) {
-        this.project = helloWorldProject;
-        this.context = ctx;
-    }
+  public SourceCompiler(Project helloWorldProject, Context ctx) {
+    this.project = helloWorldProject;
+    this.context = ctx;
+  }
 
-    public byte[] compile(){
-        return new byte[0];
-    }
-
+  public byte[] compile() {
+    return new byte[0];
+  }
 }
