@@ -2,10 +2,10 @@ package dev.secondsun.wla4j.assembler.util.analyzer;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.util.List;
 import dev.secondsun.wla4j.assembler.analyzer.Context;
 import dev.secondsun.wla4j.assembler.analyzer.SourceAnalyzer;
 import dev.secondsun.wla4j.assembler.util.TestUtils;
+import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -15,10 +15,7 @@ public class BankSizeTest {
   @DisplayName("There can be only one")
   public void testOnlyOneRomBankSize() {
 
-    var mainS = """
-          .ROMBANKSIZE $8000
-          .ROMBANKSIZE $8000
-        """;
+    var mainS = ".ROMBANKSIZE $8000\n" + ".ROMBANKSIZE $8000";
     var parser = TestUtils.asParser(mainS);
     var nodes = List.of(parser.nextNode(), parser.nextNode());
 
@@ -32,10 +29,8 @@ public class BankSizeTest {
   @DisplayName("There can be only one")
   public void testOnlyOneBankSize() {
 
-    var mainS = """
-          .BANKSIZE $8000
-          .BANKSIZE $8000
-        """;
+    var mainS = ".BANKSIZE $8000\n" + ".BANKSIZE $8000\n";
+
     var parser = TestUtils.asParser(mainS);
     var nodes = List.of(parser.nextNode(), parser.nextNode());
 
@@ -43,9 +38,5 @@ public class BankSizeTest {
     var errors = checker.analyzeProject("main.s", nodes);
     assertEquals(1, errors.size());
     assertEquals(2, errors.get(0).getSourceToken().getPosition().beginLine);
-
-
   }
-
-
 }
