@@ -1,12 +1,11 @@
 package dev.secondsun.wla4j.assembler.util.analyzer;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.fail;
 
-import java.util.List;
 import dev.secondsun.wla4j.assembler.analyzer.Context;
 import dev.secondsun.wla4j.assembler.analyzer.SourceAnalyzer;
 import dev.secondsun.wla4j.assembler.util.TestUtils;
+import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -15,14 +14,13 @@ public class SnesHeaderAnalyzerTest {
   @Test
   @DisplayName("SNES Header Must be defined only once")
   public void snesHeaderTestOnlyDefinedOnce() {
-    var source = """
-                .SNESHEADER
-                  ID "Test"
-                .ENDSNES
-                .SNESHEADER
-                ID "Test"
-                .ENDSNES
-                """;
+    var source =
+        ".SNESHEADER\n"
+            + "                  ID \"Test\"\n"
+            + "                .ENDSNES\n"
+            + "                .SNESHEADER\n"
+            + "                ID \"Test\"\n"
+            + "                .ENDSNES";
 
     var parser = TestUtils.asParser(source);
     var nodes = List.of(parser.nextNode(), parser.nextNode());
@@ -37,11 +35,7 @@ public class SnesHeaderAnalyzerTest {
   @Test
   @DisplayName("id Must Be Between One And Four Characters")
   public void idMustBeBetweenOneAndFourCharacters() {
-    var source = """
-                .SNESHEADER
-                  ID "SNES12"
-                .ENDSNES
-                """;
+    var source = ".SNESHEADER\n" + "                  ID \"SNES12\"\n" + "                .ENDSNES";
 
     var parser = TestUtils.asParser(source);
     var nodes = List.of(parser.nextNode());
@@ -56,11 +50,11 @@ public class SnesHeaderAnalyzerTest {
   @Test
   @DisplayName("id Must Be Between One And Four Characters")
   public void idMustBeBetweenOneAndFourCharacters2() {
-    var source = """
-                .SNESHEADER
-                  ID ""
-                .ENDSNES
-                """;
+    var source =
+        "\n"
+            + "                .SNESHEADER\n"
+            + "                  ID \"\"\n"
+            + "                .ENDSNES";
 
     var parser = TestUtils.asParser(source);
     var nodes = List.of(parser.nextNode());
@@ -72,16 +66,14 @@ public class SnesHeaderAnalyzerTest {
     assertEquals(1, errors.size());
   }
 
-
   @Test
   @DisplayName("Name Must Be Between 1 And 21 Characters")
   public void nameMustBeBetweenOneAnd21Characters() {
-    var source = """
-                .SNESHEADER
-                  ID "SNES"
-                  NAME "1234567890123456789012"
-                .ENDSNES
-                """;
+    var source =
+        " .SNESHEADER\n"
+            + "                  ID \"SNES\"\n"
+            + "                  NAME \"1234567890123456789012\"\n"
+            + "                .ENDSNES";
 
     var parser = TestUtils.asParser(source);
     var nodes = List.of(parser.nextNode());
@@ -96,12 +88,11 @@ public class SnesHeaderAnalyzerTest {
   @Test
   @DisplayName("Name  Must Be Between 1 And 21 Characters")
   public void nameMustBeBetweenOneAnd21Characters2() {
-    var source = """
-                .SNESHEADER
-                  ID "SNES"
-                  NAME ""
-                .ENDSNES
-                """;
+    var source =
+        ".SNESHEADER\n"
+            + "                  ID \"SNES\"\n"
+            + "                  NAME \"\"\n"
+            + "                .ENDSNES";
 
     var parser = TestUtils.asParser(source);
     var nodes = List.of(parser.nextNode());
@@ -116,14 +107,13 @@ public class SnesHeaderAnalyzerTest {
   @Test
   @DisplayName("ROM Mode must be set only once")
   public void setRomModeOnlyOnce() {
-    var source = """
-                .SNESHEADER
-                  ID "SNES"
-                  NAME "This Name"
-                  LOROM
-                  HIROM
-                .ENDSNES
-                """;
+    var source =
+        ".SNESHEADER\n"
+            + "                  ID \"SNES\"\n"
+            + "                  NAME \"This Name\"\n"
+            + "                  LOROM\n"
+            + "                  HIROM\n"
+            + "                .ENDSNES";
 
     var parser = TestUtils.asParser(source);
     var nodes = List.of(parser.nextNode());
@@ -138,15 +128,14 @@ public class SnesHeaderAnalyzerTest {
   @Test
   @DisplayName("Cartridge Type must be set only once")
   public void setCartridgeTypeOnlyOnce() {
-    var source = """
-                .SNESHEADER
-                  ID "SNES"
-                  NAME "This Name"
-                  HIROM
-                  CARTRIDGETYPE $FF
-                  CARTRIDGETYPE $01
-                .ENDSNES
-                """;
+    var source =
+        ".SNESHEADER\n"
+            + "                  ID \"SNES\"\n"
+            + "                  NAME \"This Name\"\n"
+            + "                  HIROM\n"
+            + "                  CARTRIDGETYPE $FF\n"
+            + "                  CARTRIDGETYPE $01\n"
+            + "                .ENDSNES";
 
     var parser = TestUtils.asParser(source);
     var nodes = List.of(parser.nextNode());
@@ -157,6 +146,4 @@ public class SnesHeaderAnalyzerTest {
     var errors = checker.analyzeProject("main.s", nodes);
     assertEquals(1, errors.size());
   }
-
-
 }
